requirements.txt
aiofiles==0.8.0
aiosqlite==0.17.0
asyncpg==0.26.0
distlib==0.3.6
filelock==3.8.0
httptools==0.4.0
iso8601==1.0.2
isort==5.10.1
multidict==6.0.2
pbr==5.10.0
platformdirs==2.5.2
pycryptodome==3.15.0
pydantic==1.9.2
PyJWT==2.4.0
pypika-tortoise==0.1.6
python-dotenv==0.21.0
pytz==2022.2.1
PyYAML==6.0
sanic==22.6.2
sanic-ext==22.6.3
sanic-jwt==1.8.0
sanic-routing==22.3.0
stevedore==4.0.0
tortoise-orm==0.19.2
typing_extensions==4.3.0
ujson==5.4.0
uvloop==0.16.0
virtualenv==20.16.3
virtualenv-clone==0.5.7
websockets==10.3
